+++ CHANGELOG.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
- - 
+ - Support for mpeg video files
 
 ### Fixed